Pushed down membmers of AgendaItemImpl into RuleAgendaItem and deleted AgendaItemImpl
